better code injection to pyclesperanto

+++ src/test/java/net/clesperanto/python/codegenerator/PyClEsperantoDocumentationInjector.java
@@ -1,10 +1,12 @@
 package net.clesperanto.python.codegenerator;
 
+import jnr.ffi.Struct;
 import net.haesleinhuepf.clij.macro.CLIJMacroPlugin;
 import net.haesleinhuepf.clij.macro.CLIJMacroPluginService;
 import net.haesleinhuepf.clij.macro.documentation.OffersDocumentation;
 import net.haesleinhuepf.clij2.utilities.HasAuthor;
 import net.haesleinhuepf.clij2.utilities.HasLicense;
+import net.haesleinhuepf.clijx.codegenerator.DocumentationGenerator;
 import org.python.modules._py_compile;
 import org.scijava.Context;
 
@@ -25,6 +27,7 @@ public class PyClEsperantoDocumentationInjector {
             "_tier2",
             "_tier3",
             "_tier4",
+            "_tier5",
             "_tier9"
     };
 
@@ -49,6 +52,8 @@ public class PyClEsperantoDocumentationInjector {
                         System.out.println("ex " + clij2macroMethodName);
 
                         String content = new String(Files.readAllBytes(Paths.get(filename)));
+                        String orignal_content = content;
+
                         String[] temp = content.split("\"\"\"");
                         System.out.println(temp.length);
                         if (temp.length > 2) {
@@ -63,48 +68,88 @@ public class PyClEsperantoDocumentationInjector {
                             boolean numpy_style_parameters_contained = false;
                             if (plugin instanceof OffersDocumentation) {
                                 String desc = ((OffersDocumentation) plugin).getDescription();
-                                documentation.append(desc.replace("\n", "\n    ") + "\n\n");
-                                numpy_style_parameters_contained = plugin.getParameterHelpText().contains("Parameters\n----");
+                                desc = handleLineBreaks(desc, 72);
+                                desc = desc + "\n\n";
+                                numpy_style_parameters_contained = ((OffersDocumentation) plugin).getDescription().contains("Parameters\n----");
+                                documentation.append(desc);
                             }
                             if (plugin instanceof HasAuthor) {
-                                documentation.append("    Author(s): " + ((HasAuthor) plugin).getAuthorName() + "\n\n");
+                                documentation.append("Author(s): " + ((HasAuthor) plugin).getAuthorName() + "\n\n");
                             }
                             if (plugin instanceof HasLicense) {
-                                documentation.append("    License: " + ((HasLicense) plugin).getLicense() + "\n\n");
+                                documentation.append("License: " + ((HasLicense) plugin).getLicense() + "\n\n");
                             }
-                            if (plugin instanceof OffersDocumentation) {
-                                documentation.append("    Available for: " + ((OffersDocumentation) plugin).getAvailableForDimensions() + "\n\n");
-                            }
+                            //if (plugin instanceof OffersDocumentation) {
+                            //    documentation.append("    Available for: " + ((OffersDocumentation) plugin).getAvailableForDimensions() + "\n\n");
+                            //}
 
                             String[] parameters = plugin.getParameterHelpText().split(",");
 
                             if (!numpy_style_parameters_contained) {
-                                documentation.append("    Parameters\n" +
-                                        "    ----------\n");
-                                for (String param : plugin.getParameterHelpText().split(",")) {
+                                documentation.append("Parameters\n" +
+                                        "----------\n");
+                                for (String param : parameters) {
+                                    String[] temp2 = param.trim().split(" ");
                                     documentation.append(
-                                        "    " + param + "\n");
+                                        temp2[temp2.length - 1] + " : " + temp2[temp2.length - 2] + "\n");
+                                }
+                                documentation.append("\n");
+                            }
+
+                            String parameters_call = "";
+                            String returns = null;
+                            for (String parameter : parameters) {
+                                String[] temp2 = parameter.trim().split(" ");
+
+                                if (parameters_call.length() > 0) {
+                                    parameters_call = parameters_call + ", ";
                                 }
 
-                                documentation.append("\n    Returns\n" +
-                                        "    -------\n\n    ");
+                                if (temp2[0].contains("ByRef")) {
+                                    returns = temp2[temp2.length - 1];
+                                }
+                                parameters_call = parameters_call + temp2[temp2.length - 1];
+                            }
+
+                            if (returns != null) {
+                                documentation.append("\nReturns\n" +
+                                        "-------\n" +
+                                        returns + "\n\n");
+                            }
+
+                            String example = DocumentationGenerator.generateClePythonCode(plugin.getClass(), clij2macroMethodName, "", parameters_call, "");
+                            if (example != null) {
+
+                                example = ("\n" + example)
+                                        .replace("<pre class=\"highlight\">", "")
+                                        .replace("</pre>", "");
+                                while (example.contains("\n\n")) {
+                                    example = example.replace("\n\n", "\n");
+                                }
+                                example = example.substring(0, example.length() - 1 ).replace("\n", "\n>>> ");
+
+                                documentation.append(
+                                        "Examples\n" +
+                                                "--------" +
+                                                example +
+                                                "\n\n");
                             }
 
                             documentation.append(
-                                    "    Examples\n" +
-                                    "    --------\n" +
-                                    "    \n" +
-                                    "    ");
-
-                            documentation.append(
-                                    "    References\n" +
-                                    "    ----------\n" +
-                                    "    .. [1] https://clij.github.io/clij2-docs/reference_" + clij2macroMethodName.replace("CLIJ2_", "") +
-                                    "    \n\n"
+                                    "References\n" +
+                                    "----------\n" +
+                                    ".. [1] https://clij.github.io/clij2-docs/reference_" + clij2macroMethodName.replace("CLIJ2_", "") +
+                                    "\n"
                             );
 
+                            String temp2 = documentation.toString();
 
-                            temp[1] = documentation.toString();
+                            while (temp2.contains("\n\n\n")) {
+                                temp2 = temp2.replace("\n\n\n", "\n\n");
+                            }
+                            temp2 = temp2.replace("\n", "\n    ");
+
+                            temp[1] = temp2;
 
                             content = "";
                             for (int i = 0; i < temp.length; i++) {
@@ -113,8 +158,11 @@ public class PyClEsperantoDocumentationInjector {
                                 }
                                 content = content + temp[i];
                             }
+                            content = content.replace("    >>>     \n","");
 
-                            Files.write(Paths.get(filename), content.getBytes());
+                            if (content.replace("\r\n", "\n").compareTo(orignal_content.replace("\r\n", "\n")) != 0) {
+                                Files.write(Paths.get(filename), content.getBytes());
+                            }
                         }
                     }
                 }
@@ -151,6 +199,34 @@ public class PyClEsperantoDocumentationInjector {
 
     }
 
+    private static String handleLineBreaks(String desc, int max_length) {
+        StringBuilder output = new StringBuilder();
+        desc = desc.replace("\r\n", "\n");
+        desc = desc.replace("\t", "    ");
+
+        //desc = desc.replace("\n", " \n ");
+        int counter = 0;
+        for (String element : desc.split(" ")) {
+            if (element.contains("\n")) {
+                counter = 0;
+                output.append(element + " ");
+                continue;
+            }
+
+            if (counter + element.length() >= max_length) {
+                counter = 0;
+                output.append("\n");
+            }
+            output.append(element + " ");
+            counter = counter + element.length() + 1;
+        }
+        String temp = output.toString();
+        while (temp.substring(temp.length()).compareTo(" ") == 0) {
+            temp = temp.substring(0, temp.length() - 1 );
+        }
+        return temp;
+    }
+
     private static String niceName(String name) {
         String result = "";
 
@@ -168,6 +244,7 @@ public class PyClEsperantoDocumentationInjector {
         result = result.replace("paste3_d", "paste");
         result = result.replace("minimum3_d", "minimum");
         result = result.replace("maximum3_d", "maximum");
+        result = result.replace("difference_of_gaussian3_d", "difference_of_gaussian");
         result = result.replace("mean3_d", "mean");
         result = result.replace("blur3_d", "blur");
         result = result.replace("flip3_d", "flip");
@@ -182,7 +259,7 @@ public class PyClEsperantoDocumentationInjector {
         result = result.replace("xgreater", "x_greater");
         result = result.replace("xsmaller", "x_smaller");
         result = result.replace("xequals", "x_equals");
-        result = result.replace("_pixelindex", "_pixel_index");
+        result = result.replace("_pixel_index", "_pixelindex");
